refactor: DRY up UI integration tests
- Extract setupWorkspaceView and cleanupView helpers into helpers.ts
- Consolidate waitForDirtyStatus/waitForCleanStatus with shared waitForGitStatus
- Add waitForAheadStatus helper for ahead count assertions
- Reuse setupWorkspaceView in reviewRefresh.integration.test.ts
- Replace repeated cleanup blocks with cleanupView calls

Net reduction of ~55 lines while improving test consistency.
